storeID uint64

+++ ecwid.go
@@ -18,7 +18,7 @@ type (
 )
 
 // New method creates a new ecwid client based on resty.Client
-func New(storeID uint, token string) *Client {
+func New(storeID uint64, token string) *Client {
 	client := resty.New().SetHostURL(fmt.Sprintf(endpoint, storeID)).SetQueryParam("token", token)
 	return &Client{
 		client,
